requirements/docs.txt
Pillow~=9.5
Sphinx~=7.4
sphinx_gallery~=0.14
sphinx_rtd_theme~=2.0
sphinx_substitution_extensions==2024.8.6
setuptools-git-versioning==2.0.0
autodoc_pydantic==2.2.0